chore: create memhog
Creates an application that consumes memory incremntally to exercise
resource limits. The program allocates memory incrementally and exits
when it receives a signal or when it reaches a specified allocation
limit. A pod manifest is added to run the program with a memory limit so
k8s will terminate it for us as a baseline for non-DinD behaviour.

Manifests for metrics server are also added to enable `kubectl top`.